Enable MHL in defconfig, /proc/config.gz support, make default config rainbow
MHL is still not working - stuff appears in dmesg now but no video.  Unknown if this is remaining
kernel work (broken G2D driver?) or something missing in userland.  (TW launcher had to be active
in Froyo.)  However this does not seem to be negatively affecting it.

/proc/config.gz support - well, that should always be enabled

Default config rainbow - norainbow is only for CWM, it's the "odd kernel out" for the time being